Added a bunch of features, see description. Fixed bugs
Everything is now optional, clientId is now able to be refreshed while the client is running. Images are now able to be disabled, and can be a array (so they can be randomized and chosen at random). Timestamp able to be refreshed while app running. Quotes are now able to be static (you can just have one). See the README.md file for more info (if there is more. Is there?)